Fix false positive on request validation.
