+++ readme.md
@@ -106,5 +106,5 @@ Lints:
 - File should end with a new line `NewLineAtEndOfFile`
 - No /** @var ClassName $var */ inline docs `NoInlineVarDocs`
 - There should be no unused imports `NoUnusedImports`
-- No `dd()` usages `NoDd`
-- Use `request()->validate(...)` helper function over `$this->validate(...)` in controllers `RequestHelperValidation`
+- There should be no calls to `dd()` `NoDd`
+- Use `request()->validate(...)` helper function or extract a FormRequest instead of using `$this->validate(...)` in controllers `RequestValidation`
